Use generic picker for statistic picker
